SONARJAVA-5509 Refactor of S1481 test samples to enable testing of multiple java versions
