+++ java-checks/src/test/java/org/sonar/java/checks/unused/UnusedLocalVariableCheckTest.java
@@ -16,6 +16,7 @@
  */
 package org.sonar.java.checks.unused;
 
+import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.CheckVerifier;
 import org.sonar.java.checks.verifier.TestUtils;
@@ -23,12 +24,15 @@ import org.sonar.java.checks.verifier.TestUtils;
 class UnusedLocalVariableCheckTest {
 
   @Test
-  void test() {
-    CheckVerifier.newVerifier()
-      .onFile(TestUtils.mainCodeSourcesPath("checks/unused/UnusedLocalVariableCheck.java"))
-      .withCheck(new UnusedLocalVariableCheck())
-      .withJavaVersion(22)
-      .verifyIssues();
+  void test_all_versions() {
+    List<Integer> versions = List.of(20, 21, 22);
+    for(int version : versions) {
+      CheckVerifier.newVerifier()
+        .onFile(TestUtils.mainCodeSourcesPath("checks/unused/UnusedLocalVariableCheck.java"))
+        .withCheck(new UnusedLocalVariableCheck())
+        .withJavaVersion(version)
+        .verifyIssues();
+    }
   }
 
   @Test
@@ -40,23 +44,12 @@ class UnusedLocalVariableCheckTest {
       .verifyIssues();
   }
 
-  /** Check that issue that can only be acted upon with Java 22 are not raised for earlier versions. */
   @Test
   void test_java21() {
     CheckVerifier.newVerifier()
-      .onFile(TestUtils.mainCodeSourcesPath("checks/unused/UnusedLocalVariableCheck_java22.java"))
+      .onFile(TestUtils.mainCodeSourcesPath("checks/unused/UnusedLocalVariableCheck_java21.java"))
       .withCheck(new UnusedLocalVariableCheck())
       .withJavaVersion(21)
-      .verifyNoIssues();
-  }
-
-  /** Verify no regression older versions of java */
-  @Test
-  void test_java17(){
-    CheckVerifier.newVerifier()
-      .onFile(TestUtils.mainCodeSourcesPath("checks/unused/UnusedLocalVariableCheck_java17.java"))
-      .withCheck(new UnusedLocalVariableCheck())
-      .withJavaVersion(17)
       .verifyIssues();
   }
